requirements.txt
Flask
Pillow
requests
google-cloud-storage